requirements.txt
limacharlie
cherrypy
pyyaml